This allows our k quant mat vec shaders to process multiple rows at a time just like `mul_mat_vec.comp`. It's way faster now and Q4_K_S is catching up to IQ4_NL and Q4_0 on my RX 470.

At this point we might want to consider merging the separate k quant files into `mul_mat_vec.comp` as they're reusing quite a bit of code, and maybe do some templating using ifdefs to choose the correct dequantization function. That's better left to another PR though.

PR:
| model                          |       size |     params | backend    | ngl | threads |   main_gpu |    sm |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ---------: | ----: | ------------: | -------------------: |
| llama 8B Q2_K - Medium         |   2.95 GiB |     8.03 B | Vulkan     | 100 |       8 |          1 |  none |         tg128 |         21.88 ± 0.00 |
| llama 8B Q3_K - Medium         |   3.74 GiB |     8.03 B | Vulkan     | 100 |       8 |          1 |  none |         tg128 |         18.89 ± 0.04 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     | 100 |       8 |          1 |  none |         tg128 |         27.12 ± 0.12 |
| llama 8B Q5_K - Small          |   5.21 GiB |     8.03 B | Vulkan     | 100 |       8 |          1 |  none |         tg128 |         22.55 ± 0.00 |
| llama 7B Q6_K                  |   5.53 GiB |     7.24 B | Vulkan     | 100 |       8 |          1 |  none |         tg128 |         20.39 ± 0.00 |

```
  MUL_MAT(type_a=q2_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   4260 runs -   24
2.40 us/run - 117.44 MFLOP/run - 484.49 GFLOPS
  MUL_MAT(type_a=q3_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   2556 runs -   44
7.88 us/run - 117.44 MFLOP/run - 262.22 GFLOPS
  MUL_MAT(type_a=q4_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   4260 runs -   23
8.10 us/run - 117.44 MFLOP/run - 493.23 GFLOPS
  MUL_MAT(type_a=q5_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   3408 runs -   31
5.13 us/run - 117.44 MFLOP/run - 372.68 GFLOPS
  MUL_MAT(type_a=q6_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   3408 runs -   36
8.04 us/run - 117.44 MFLOP/run - 319.10 GFLOPS
```

Master:
| model                          |       size |     params | backend    | ngl | threads |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ------------: | -------------------: |
| llama 8B Q2_K - Medium         |   2.95 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |         17.66 ± 0.07 |
| llama 8B Q3_K - Medium         |   3.74 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |         15.74 ± 0.02 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |         20.58 ± 0.00 |
| llama 8B Q5_K - Small          |   5.21 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |         16.04 ± 0.01 |
| llama 7B Q6_K                  |   5.53 GiB |     7.24 B | Vulkan     | 100 |       8 |  tg128  |  17.57 ± 0.06  |

```
  MUL_MAT(type_a=q2_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   3408 runs -   39
1.03 us/run - 117.44 MFLOP/run - 300.33 GFLOPS
  MUL_MAT(type_a=q3_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   2556 runs -   52
8.63 us/run - 117.44 MFLOP/run - 222.16 GFLOPS
  MUL_MAT(type_a=q4_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   3408 runs -   37
4.13 us/run - 117.44 MFLOP/run - 313.90 GFLOPS
  MUL_MAT(type_a=q5_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   2556 runs -   47
2.77 us/run - 117.44 MFLOP/run - 248.41 GFLOPS
  MUL_MAT(type_a=q6_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   2556 runs -   46
1.75 us/run - 117.44 MFLOP/run - 254.34 GFLOPS
```

The number of rows used was chosen for my card and may need tuning for different architectures.