@netrunnereve 2024-12-19
> Edit: Also looks good on AMD RX 6800 XT.

Does that mean it works best with two rows per shader?