@netrunnereve 2024-12-19
This should be it I think:

Default: 2 rows for old quants, 1 row for Q8_0 and K quants
AMD GCN: 4 rows for old quants, 2 rows for Q8_0, 4 rows for K quants
AMD RDNA: 2 rows for old quants, 1 row for Q8_0, 2 rows for K quants
Intel: 4 rows for old quants, 2 rows for Q8_0, 2 rows for K quants