@0cc4m 2024-12-21
I see a significant drop in performance on Nvidia RTX 3090 in tg for a Q4_K_S 8B model:

```
Master:
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         88.64 ± 0.17 |
PR:
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         77.01 ± 0.43 |
With rm_kq = 2:
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         99.12 ± 2.98 |
```

This doesn't match with the results from @jeffbolznv. Any theories?